+++ Src/Commands/ExportDataFrameCommand.cs
@@ -1,9 +1,7 @@
 ﻿
 using Microsoft.Data.Analysis;
-using System;
-using System.IO;
-using System.Management.Automation;
 using Parquet;
+using System.Management.Automation;
 
 namespace PSDataFrame.Commands;
 
@@ -15,14 +13,13 @@ public class ExportDataFrameCommand : BaseImportExportCommand
     public DataFrame DataFrame { get; set; }
 
     [Parameter(ParameterSetName = PsnPath, Position = 1, Mandatory = true)]
-    [Parameter(ParameterSetName = PsnParquet, Position = 1, Mandatory = true)]
     public string Path { get; set; }
+
+    [Parameter(ParameterSetName = PsnParquet, Mandatory = true)]
+    public string ParquetPath { get; set; }
 
     [Parameter(ParameterSetName = PsnString, Mandatory = true)]
     public SwitchParameter String { get; set; }
-
-    [Parameter(ParameterSetName = PsnParquet, Mandatory = true)]
-    public SwitchParameter Parquet { get; set; }
 
     protected override void BeginProcessing()
     {
@@ -61,10 +58,9 @@ public class ExportDataFrameCommand : BaseImportExportCommand
 
         if (ParameterSetName == PsnParquet)
         {
-            using FileStream parquetStream = new FileStream(GetUnresolvedProviderPathFromPSPath(Path), FileMode.Create, FileAccess.Write, FileShare.None);
+            using var parquetStream = new FileStream(GetUnresolvedProviderPathFromPSPath(ParquetPath), FileMode.Create, FileAccess.Write, FileShare.None);
 
-            //they only have async extension method
-            DataFrame.WriteAsync(parquetStream).Wait();
+            DataFrame.WriteAsync(parquetStream).GetAwaiter().GetResult();
             return;
         }