+++ Release-Notes.md
@@ -1,5 +1,13 @@
 # DataFrame Release Notes
 
+## v0.3.0
+
+**Breaking change** Requires PowerShell 7.4+
+
+Support Parquet files export and import.
+
+Microsoft.Data.Analysis 0.22.2
+
 ## v0.2.3
 
 Microsoft.Data.Analysis 0.22.0